requirements.txt
# Core dependencies
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
scipy>=1.7.0
pymongo>=4.0.0
python-dotenv>=0.19.0

# Web server (FastAPI)
fastapi>=0.95.0
uvicorn>=0.21.1
python-multipart>=0.0.6
pydantic>=1.10.7
aiofiles>=23.1.0
aiohttp>=3.8.0

# Gmail API
google-auth>=2.16.0
google-auth-oauthlib>=0.8.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.70.0

# Image processing
pillow>=8.3.0
pytesseract>=0.3.8
opencv-python>=4.5.0

# PDF processing
PyPDF2>=2.0.0
pdfminer.six>=20200720

# NLP and text processing
nltk>=3.6.0
rapidfuzz>=3.0.0

# Utilities
tqdm>=4.62.0
requests>=2.26.0
pytz>=2021.1

# Testing
pytest>=6.2.5
python_dotenv